+++ src/main/java/insane96mcp/insanesurvivaloverhaul/data/generator/ISOItemTagsProvider.java
@@ -4,6 +4,7 @@ import insane96mcp.insanesurvivaloverhaul.InsaneSO;
 import insane96mcp.insanesurvivaloverhaul.module.death.respawn.EchoPillar;
 import insane96mcp.insanesurvivaloverhaul.module.farming.crops.Crops;
 import insane96mcp.insanesurvivaloverhaul.module.hungerhealth.fooddrinks.FoodDrinks;
+import insane96mcp.insanesurvivaloverhaul.module.items.StackSizes;
 import insane96mcp.insanesurvivaloverhaul.module.items.UnvanishableItems;
 import insane96mcp.insanesurvivaloverhaul.module.items.copper.CopperEquipment;
 import insane96mcp.insanesurvivaloverhaul.module.mining.beegoreveins.BigOreVeins;
@@ -118,6 +119,9 @@ public class ISOItemTagsProvider extends ItemTagsProvider {
 
         tag(BigOreVeins.ORE_ROCKS_ITEM)
                 .add(BigOreVeins.COPPER_ORE_ROCK.item().get(), BigOreVeins.IRON_ORE_ROCK.item().get(), BigOreVeins.GOLD_ORE_ROCK.item().get());
+
+        tag(StackSizes.NO_STACK_SIZE_CHANGES)
+                .add(Items.GOLDEN_CARROT);
     }
 
     public static TagKey<Item> create(String tagName) {